Extract topEntity name from .topentity file
