+++ CLaSH.hs
@@ -1,4 +1,5 @@
 import CLaSH.Driver
+import CLaSH.Driver.TopWrapper
 import CLaSH.Rewrite.Types
 import CLaSH.GHC.Evaluator
 import CLaSH.GHC.GenerateBindings
@@ -22,9 +23,10 @@ doHDL :: Backend s
        -> IO ()
 doHDL b src = do
   pd      <- primDir b
-  primMap <- generatePrimMap [pd]
+  primMap <- generatePrimMap [pd,"."]
+  topEntM <- generateTopEnt "./fir.topentity"
   (bindingsMap,tcm) <- generateBindings primMap src Nothing
-  generateHDL bindingsMap (Just b) primMap tcm ghcTypeToHWType reduceConstant DebugNone
+  generateHDL bindingsMap (Just b) primMap tcm ghcTypeToHWType reduceConstant topEntM DebugNone
 
 main :: IO ()
 main = genVHDL "./examples/FIR.hs"
