+++ clash-lib/src/CLaSH/Primitives/Util.hs
@@ -1,9 +1,7 @@
 -- | Utility functions to generate Primitives
 module CLaSH.Primitives.Util where
 
-import           Data.Aeson             (FromJSON, Result (..), fromJSON, json)
-import qualified Data.Attoparsec.Lazy   as L
-import           Data.ByteString.Lazy   (ByteString)
+import           Data.Aeson.Extra       (decodeAndReport)
 import qualified Data.ByteString.Lazy   as LZ
 import qualified Data.HashMap.Lazy      as HashMap
 import           Data.List              (isSuffixOf)
@@ -36,15 +34,3 @@ generatePrimMap filePaths = do
   let primMap = HashMap.fromList $ zip (map name primitives) primitives
 
   return primMap
-
--- | Parse a ByteString according to the given JSON template. Prints failures
--- on @stdout@, and returns 'Nothing' if parsing fails.
-decodeAndReport :: (FromJSON a)
-                => ByteString -- ^ Bytestring to parse
-                -> Maybe a
-decodeAndReport s =
-  case L.parse json s of
-    L.Done _ v -> case fromJSON v of
-                    Success a -> Just a
-                    Error msg -> traceIf True msg Nothing
-    L.Fail _ _ msg -> traceIf True msg Nothing
